Link README to documentation
Move the build definition description into the documentation. Retain a
basic example in the README.

Add docs and example directories to the directory structure
description, and add a sub heading to group the Go code directories.

Reference the module types from the build definition documentation.

Revise index.md to point to the main descriptive information first,
and then list the pages with reference information. Also fix module
links which should be pointing to *.md files.

Change-Id: Id65e021bed10af02b9cf1e8aa74cf7259c72be57
Signed-off-by: David Kilroy <david.kilroy@arm.com>