Use BlueprintEmbed for feature properties
Go's reflect.StructOf() implementation imposes a limit on the number
of fields in dynamically-created types. This is because it tries to
generate a name for the type based on printing out the entire type's
fields, including embedded structs. This name is limited to 2^16-1
bytes (see newName in reflect/type.go).

This causes problems in Bob when there are lots of features - we
start to approach that limit when multiplied by the size of
representing e.g. the BuildProps struct.

Work around this by not embedding each properties structure directly
inside each feature. Instead, each feature now points to another
BlueprintEmbed interface{}, which is populated after creating the
features type. This shrinks the type name representation because
each feature now only contains a single field, instead of an entire
properties structure.

This change requires an update of the tests, because there is now an
extra level of indirection between the Features type and each
feature's properties. To handle this, make the feature name argument
of injectData() explicit, so it can get through the second layer of
BlueprintEmbed before trying to find the target value.

Change-Id: I8bfd9afa4cb89440ffa4d28ed5f765968be18c65