Fix creation of embedded fields with reflection
Creating a struct with an anonymous field (e.g. another embedded
struct) differs between Go versions.

In Go 1.8, the 'Anonymous' field is ignored, and the field's
anonymity is indicated by its 'Name' property being empty.

In Go 1.10, the 'Anonymous' field is _not_ ignored, and the name must
_not_ be empty (otherwise reflect.StructOf will panic).

This means that our current implementation of coalesceTypes only
works on Go 1.10.

Fix this by clearing 'name' on anonymous fields - but only after
running a test to find out which behavior the current implementation
uses. Rather than doing a version check, a function is added which
tries to create a new type using reflection, then checks if the
'Anonymous' property was respected.

Change-Id: Icf8ee2addb2e0a8b7bf9373edd4a7b0dad1d49fd