Fix generated libraries on Android O and later
From Android O, the Android build system checks that libraries are
allowed to link to each other. In order for this to work, each library
needs to declare its link_type.

For our generated libraries to work we need to do the same. To support
this we output a similar fragment of code to that in
build/core/make/binary.mk to setup the appropriate variables and
files.

Android O and P have different implementations of this, so output an
implementation for each and assume future Android releases will be
similar to Android P.

Change-Id: I972f5cc9f75634fc6b8d70cb668ad32640f144dd
Signed-off-by: David Kilroy <david.kilroy@arm.com>