Fix C++ cross-compilation with Clang
Cross-compilation with Clang is broken for two reasons:
 - It cannot find the libstdc++ headers
 - It cannot _always_ find a linker compatible with the target
   (although if both compilers are installed as system packages this
   may work automatically)

Fix these by (respectively):
 - Passing the --gcc-toolchain option to Clang
 - Adding the GCC toolchain's linker to Clang's binary search path
   (on system-installed packages these directories will be
   duplicates, so are removed automatically)

This is tested with:
 - Standalone Clang and standalone GCC
 - System-installed Clang and standalone GCC

It currently does *not* work with system-installed cross-GCC
toolchains. This is because GCC's sysroot is '/' - when that is used
to generate C++ header paths in newToolchainClangCross(), we end up
with non-existent '/../include/...' paths. This will be fixed in a
subsequent commit.

Change-Id: I943c9a01a7065bf5c982c3585f3735a0296ff20d
Signed-off-by: Chris Diamand <chris.diamand@arm.com>