Move shared Clang toolchain code into new type
Create a 'common' type for toolchainClang[Cross|Native], so that the
common code can be shared more easily.

Clang's common type now includes a reference to its corresponding GNU
toolchain as an interface, rather than a concrete 'Native' or 'Cross'
type, which means that the getArchiver et al methods, which delegate
to the GNU toolchain, can be shared between cross and native objects.

Change-Id: I4040a4351c830c3850852ca706859150e1f70b7f
Signed-off-by: Chris Diamand <chris.diamand@arm.com>