Improve libstdc++ header path discovery
GCC toolchains ship with their own C++ headers, which are used in
Clang builds in order to link with libstdc++.

Currently, we use a path relative to the output of `gcc
-print-sysroot`. However, because it contains '..', it becomes
invalid when the GCC toolchain is part of the system installation,
and therefore has '/' as its sysroot.

Furthermore, using the TARGET_SYSROOT to find these headers means we
can't use an _actual_ sysroot with Clang - it always has to refer to
a GCC installation.

Fortunately, this can be fixed: We can reliably determine the C++
header path of any pre-packaged GCC installation, because these
headers consistently live in the '{target}/include/c++/{version}' and
'{target}/include/c++/{version}/{target}' subdirectories.

When the compiler is not a cross-compiler, the first '{target}'
subdirectory is omitted.

This commit should not change the user-visible behaviour in any way,
except when using a Clang cross-compiler and system-installed GCC
toolchain, which was previously broken. Accordingly, TARGET_SYSROOT
is still filled-in in host_explore.py, even though it is no longer
needed. This will be changed in a later commit to enable using
arbitrary sysroots.

Change-Id: If4be308338c42fbbe9371192029a27d461e15373
Signed-off-by: Chris Diamand <chris.diamand@arm.com>