menuconfig: fix inputbox with integer configs
An inputbox always wants to deal with strings, so make sure we convert
any integer config values for display.

When setting the value of the config make sure we convert back to an
int, and that we won't try to test the integer against a string.

Also split MenuItem.set_bool() into MenuItem.set() and
MenuItem.clear(), so the caller is isolated from how we encode the
boolean internally.

Change-Id: I68530b2cde6fe7beda9de67f34a596a73a16fe31
Signed-off-by: David Kilroy <david.kilroy@arm.com>