Fix append() usage with toolchain flags
Toolchain flags are getting corrupted because of incorrect usage of
append(). The issue is as follows:

 (1) Toolchain initialisation generates cflags and cxxflags.
 (2) cxxflags is then merged with cflags:
         cxxflags = append(cflags, cxxflags)
 (3) In CompileObjs(), cflags is merged with conlyflags:
         ...append(cctargetflags, Properties.conlyflags)

The issue is that, because of (2), cxxflags ends up using the same
buffer as cflags. This means that the last two elements of that
buffer can be overwritten with conlyflags by (3), with the result
that conlyflags end up in cxxflags!

Fix this in two ways:
 * Swap the order of cflags and cxxflags in (2), so that the cflags
   are copied into cxxflags' buffer.
 * Instead of appending arrays then passing them to strings.Join(),
   add a new helper, utils.Join(), which can join multiple string
   arrays into a new destination string without appending any of the
   source arrays together. This prevents the overwriting in (3).

Change-Id: I652578bdda4cb7ec9cc33cccb97affa4dd44cc19
Signed-off-by: Chris Diamand <chris.diamand@arm.com>