Rename 'ARM' toolchain
Rename the 'ARM' toolchain to 'ARMCLANG', and update the help text to
refer to it as "Arm Compiler 6", matching the product's name:

  https://developer.arm.com/products/software-development-tools/compilers/arm-compiler

'ARMCLANG' is chosen because:
 - it matches the name of the compiler binary
 - it is shorter than 'ARM_COMPILER_6' or similar
 - it differentiates it from Arm Compiler 5 and below, which use
   different options and have different binary names, e.g. 'armcc'.

This commit also moves some of the code around, since it is being
changed anyway, so as to keep each toolchain in its own section
within 'toolchain.go' and the Mconfig files. AC6 support is added to
'example/Mconfig'.

Change-Id: I6de3ea361af7a7494e87c3919bc83ea032cb1346
Signed-off-by: Chris Diamand <chris.diamand@arm.com>