Allow configurable link parallelism
Permit modification of the link pool size using the
BOB_LINK_PARALLELISM environment variable, instead of hard-coding it
to num_cpus/5+1.

Because Bob now checks an environment variable, it needs to have a
mechanism to re-run Bob if the environment variable changes. This is
done by adding a hash of the current environment to the config in
generate_config_json.py, which is run at every build. If this
hash changes, config.json's will be updated with a new hash, causing
Bob to be re-run.

Change-Id: Ia93c6ca89a2379756ce0aeeedfff509d4887d62f
Signed-off-by: Chris Diamand <chris.diamand@arm.com>