Fix potential issues with misuse of append function
No issue usage.
x = append(x, a)

Problematic one.
y = append(x, a)

If the backing array of x is too small to fit all the given values a
bigger array will be allocated. The returned slice will point to the
newly allocated array.

So when append allocates new slice there is no problem in case #2.
When no new array is allocated then x is appended so y is an alias to
x..

Change-Id: I32b1c3976d0710e5c4d92c4bbe82a72e93af2b73
Signed-off-by: Dawid Drozd <dawid.drozd@arm.com>