Improve `which` implementation
The `which` implementation in `host_explore.py` did not handle the
fact that a binary can be found within the current directory, *even
if $PATH does not contain a `.`*, as long as the binary is specified
using a multi-component relative path.

For example, given the following layout:
    .
    `-- bin
        `-- gcc

`bin/gcc` can be invoked from the current directory.

Rather than actually modifying `which_binary`, this is handled by
deleting the implementation and using Python builtins instead.

Also, remove a `which_binary` invocation from the check_output
wrapper; it was only used to check if the binary didn't exist, but
that can be discovered by handling OSError or FileNotFoundError,
depending on the Python version.

Change-Id: I542e0dd60607d7faafa4cec6d8b604594ed99cd5
Signed-off-by: Chris Diamand <chris.diamand@arm.com>