Allow arbitrary sysroots
Remove explicit sysroot handling in Bob, because this was limited to
passing the output of `gcc --print-sysroot` to Clang via
`host_explore.py`. This isn't much use, as the sysroot could be
anything, and there is no way for the build system to reliably detect
this.

After this change, super-projects can setup a sysroot however they
prefer, and simply add the `--sysroot` flag to `cflags`.

Change-Id: I558a66be7dfd0f1c0224131b499995ae1947c0a4
Signed-off-by: Chris Diamand <chris.diamand@arm.com>