Separate linker flags
Allow different flags for linking vs compiling, by adding a
`getLinker` method to the toolchain interface. For now, keep the
current behaviour as much as possible, including using the C++
compiler as the actual linker (rather than calling `ld` directly, or
trying to detect whether to use the C++ or C compiler).

Change-Id: I9e753f125c556343f2286be386553e4005a2450d
Signed-off-by: Chris Diamand <chris.diamand@arm.com>