Move host exploration into `toolchain.go`
Move the code discovering parts of the GCC toolchain into
`toolchain.go`, but guard it under a config option,
`[HOST|TARGET]_CLANG_USE_GNU_LIBS`, in preparation to support
building with standalone Clang toolchains where no GCC is present.

This preserves the existing behaviour, except for one thing; the
directory containing `libgcc.a` is only added to the command line
using `-L`, instead of with `-L` and `-B`, because the `-B` appears
unnecessary.

The relevant remaining code in `host_explore.py` is simplified, by
consolidating everything related to adding libstdc++ to the rpath
under a single `if native_build:` statement.

Change-Id: Ief6e24858385f54a3f79e54ff6b1ae7352bc377d
Signed-off-by: Chris Diamand <chris.diamand@arm.com>