Fix warn_on argument parsing
The `warn_on` script was using `action="store"` for its `-w` option,
which is used to specify options to warn about. However, the action
meant that using multiple `-w` flags would not work, as the last one
would overwrite the previous ones.

Fix this by changing the action to "append".

Change-Id: I3313f6e5b1a56623dfcbe1152172ea02948baef5
Signed-off-by: Chris Diamand <chris.diamand@arm.com>