Fix uninstalled shared libraries on Android
The `out:` field tests added in `01f6e0f` revealed an issue on Android,
where multilib shared libraries cannot be linked:

    prebuilts/gcc/linux-x86/aarch64/aarch64-linux-android-4.9/aarch64-linux-android/bin/ld.gold:
        error: out/target/product/hikey960/obj/lib/libshared_output.so: incompatible target
    external/bob-build/tests/output/out.c:5: error: undefined reference to 'libshared'

This is occurring because setting `LOCAL_UNINSTALLABLE_MODULE := true`
on target shared library causes the multilib paths to conflict,
resulting in the 32-bit library overwriting the 64-bit one.

Fix this by not disabling installation of target shared libraries. This
will let Android use a default location, and the above issue does not
occur.

Add a test to ensure that shared libraries work both with and without
install groups, on both host and target.

Change-Id: Id10801fd9e14027d64c4ea9a9bf6db319f320e95
Signed-off-by: Chris Diamand <chris.diamand@arm.com>