Add external shared library module type
This new module type can be used to replace the hard-coded list of
Android libraries in `androidSharedLibs`.

One non-obvious part of this change is that we have to output the
contents of `export_shared_libs` into `LOCAL_SHARED_LIBRARIES` for
static libraries. This is because unlike the `ldlibs` field which
contained this information previously, static libraries cannot have a
non-empty `shared_libs` property.

Extend the external static library tests to cover shared libraries, too.

Change-Id: I3212c19a3411492ed8f138944192fafcd7b4f045
Signed-off-by: Chris Diamand <chris.diamand@arm.com>