Use Android's prebuilt Ninja
Include `ninja_config.mk`, which sets up the NINJA environment variable,
which contains the location of Ninja within Android's prebuilts.
Propagate this through the bootstrap process so that it is used when
building.

When NINJA is not set, default to using `ninja`, which will be found
using PATH as usual.

Change-Id: Ie9ccff0e08375ba91b5a6a70ae394fa3e886bf68
Signed-off-by: Chris Diamand <chris.diamand@arm.com>