Fix broken comment regex
Python 3.6 and later changes the behaviour of regular expressions, so
that unknown escape sequences result in errors. This highlights the
regular expression in `t_ANY_comment()`, which contains `\Z` in a
character class. `\Z` *does* correctly match the end of string
(i.e. EOF in Ply), but when used in a character class it loses its
special meaning and just becomes a normal, broken escape sequence.

Fix this by removing it. It is only required to handle the case where
the last line is a comment, but has no newline before the end-of-file.
This isn't supported in the other token types anyway, so for now there
is no need to support it in comments.

Test this by adding a formatter test, which lets us check that arbitrary
input can be scanned without issue. Since the code is being changed anyway,
with a test being added, make the formatter strip trailing whitespace
from comments.

Improve the tests' error message by printing the expected and actual
lines when a failure occurs.

Change-Id: I4dffd64ba775199c3a5c41be0adfd7a37bf399ae
Signed-off-by: Chris Diamand <chris.diamand@arm.com>