Bob bootstrap improvements
Update the common bootstrap script for Bob so that it allows source
directory, working directory and build directory to be set
independently.

Update test bootstrap to allow us to test this. There are still some
fixes that need to be done in Bob in order for the tests to work when
bootstrapped from outside the test directory. Because of this the
example bootstrap script is left as is for now too. However the check
in bob/tests/bootstrap is removed so that we're able to easily check
if things are working.

Change-Id: I31e3e6cafd004a6a0329b032d39c68129cf5440a
Signed-off-by: David Kilroy <david.kilroy@arm.com>