Prefix include directories with ${SrcDir}
This should allow the output directory to be used as the working
directory.

Modify GetExportedVariables to return local and global include path
separately, so we can prefix ${SrcDir} without calling IsAbs(). The
comment that is being removed in this function is redundant and no
longer applies to the code (this applied prior to the graph orderring
algorithm).

Change-Id: I1db6a7d431954fed32297ff5e602dec556281938
Signed-off-by: David Kilroy <david.kilroy@arm.com>