Show all output from generate_android_inc
`generate_android_inc.bash` has to be very specific about what it
outputs to stdout, because this is interpreted when it is called from
`Android.mk`.

It currently does this by swapping stdout and stderr, then writing the
'interpreted' output to stderr. However, this means that errors in
subcommands may get lost, unless we explicitly redirect the output in
everything the script invokes.

Improve this by creating a third file descriptor, which is used for the
'interpreted' output. Redirect all other output to file descriptors 1
and 2 to the script's standard error.

This means that the user can see all output, making it possible to
reliably debug when this script fails.

Change-Id: I1c05c51169221f3438d9a57ca181f900600ebc3f
Signed-off-by: Chris Diamand <chris.diamand@arm.com>