Run Android.mk generation in ANDROID_BUILD_TOP
Change the current directory when 'building' for Android (i.e.
generating `Android.mk`) from the source directory to
`ANDROID_BUILD_TOP`, which is the CWD for the actual compilation, too.

Update the `glob` function to temporarily add the source directory to
the path before calling `GlobWithDeps`. Without this, globs will
silently return zero matches, causing build failures later, because the
source files inside a glob are relative to the source directory, not the
working directory.

Change-Id: Ie8d6f357fb7b270899f0a2de7a50df0fb54a9ead
Signed-off-by: Chris Diamand <chris.diamand@arm.com>