Explain why config options were ignored
When a command-line `OPTION=value` argument is ignored, attempt to
provide an explanation, rather than just giving the cryptic "Value is
..." message.

Because this occurs during an error case, which isn't usually tested,
add a new test to ensure this keeps working. This done using the pytest
framework - in future, the other config system tests will be ported to
use that, so we won't need three different commands to test it.

Note that Pytest has two entry points - `py.test` and `pytest`. `pytest`
is newer, so is recommended in the documentation. However, Travis will
use `py.test`, because that is the only one which exists on the Ubuntu
16.04 environment it uses.

Change-Id: I723d3b7492c2002110921ffefb5bcf5c1dee2d3c
Signed-off-by: Chris Diamand <chris.diamand@arm.com>