Fix --extra-cflags argument usage in kmod_build.py call
Works:
$ kmod_build.py --extra-cflags '-Wall -Wextra'

Fails:
$ kmod_build.py --extra-cflags '-Wall'
Error: argument --extra-cflags: expected one argument

When we try to set --extra-cflags to a single option when calling
kmod_build.py, argparse ends up seeing the option (intended for make
CFLAGS) as an option to kmod_build.py. This is because shell expands the
quoting, and argparse is just seeing -Wall.

Change-Id: I4827f47cea89c5c874925a8af51c1bef28c25536
Signed-off-by: Michal Widera <michal.widera@arm.com>