Handle src==tgt in relative path
When the source and target directories are the same, relative path
should return "."

When checking for parent directories with ${DIR#PARENT}, make sure we
look for the separating '/', otherwise the prefix "build" could match
the directory "build2". Test using [[ ]] which allow us to simplify
the conditions by using shell patterns.

Add a script that tests relative_path.

Change-Id: I4374f0224b4382baf8c7988dae456ed1c26df82d
Signed-off-by: David Kilroy <david.kilroy@arm.com>