Use Android's prebuilt ply
Android ships with a copy of the `ply` module, but does not
automatically add it to `PYTHONPATH`.

When building on Android, check for the module, and add it to `sys.path`
if it exists.

To avoid duplicating the code, do this in `__init__.py` (rather than in
the two places which actually `import ply`).

Change-Id: Ibbdb2e00a56997f1fb6b38a192386808ac2d4709
Signed-off-by: Chris Diamand <chris.diamand@arm.com>