Add a glob so we can reproduce a bootstrap issue
This commit is not intended to be merged to master.

The bootstrap issue occurs when a build output directory is
re-bootstrapped and the working directory is changed.

To reproduce the issue:
 tests/bootstrap -o tests/build
 tests/build/config
 tests/build/buildme
 cd tests
 bootstrap -o build
 build/buildme

The issue is that the file tests/build/.bootstrap/build-globs.ninja
identifies the build directory relative to the old working
directory. This causes the bootstrap builder to fail as bpglob can't
be found. I think the primary builder would fix this if it could be
executed.

Change-Id: I269eef8c26352519fc0265b4904924985f8095f7
Signed-off-by: David Kilroy <david.kilroy@arm.com>