Allow working directory to change on re-bootstrap
If you re-bootstrap and change working directory, you get an error
about not being able to find bpglob.

To reproduce this the project needs to be using globs, and the
sequence should look as follows. I've included the commands for
Bob tests.

$ cd tests
$ bootstrap -o build
$ build/config
$ build/buildme
$ cd ..
$ tests/bootstrap -o tests/build
$ tests/build/buildme
Writing config JSON to ./config.json
[1/1] ./.minibootstrap/minibp .bootstrap/build.ninja
ninja:
 error: 'build/.minibootstrap/bpglob', needed by 'build/.glob/tests/resources/__/_.glob', missing and no known rule to make it

This happens because .bootstrap/build-globs.ninja contains a buildDir
variable relative to the old working directory. Fix this by
invalidating the file whenever bootstrap is called and buildDir has
changed.

I believe this file is normally updated by the primary builder, and
used by the bootstrap builder to detect when changes in the globs
should cause the primary builder to be re-run. Since the primary
builder can't run at this point, invalidating the file is the only way
to proceed.

Change-Id: Ifc7c281f980a9eadc5ac4082062e235b30b6571a
Signed-off-by: David Kilroy <david.kilroy@arm.com>