Fix generated source depfiles
Currently, the `depfile` field on generated and transformed sources is
ignored, because we do not specify the format (GCC or MSVC).

Fix this by specify the format as GCC, which causes Ninja to actually
read the generated `.d` files.

This also highlights that `depfile` was being set unconditionally,
regardless of whether the module uses depfiles. Make this conditional on
the `depfile` field not being empty.

When Depfile is specified in blueprint, the 'depfile' variable is set
for each build rule. Update the rules to use that rather than
respecify the file.

Add a generate_source test that uses `depfile`.

Add travis testing that checks that touching various sources causes
relevant files to be updated. In testing kernel module updates, add an
include of a 'kernel header'. In testing static library updates fix a
header include.

Change-Id: I8e608c4e7231aacffdb53d6549fb9c3333387991
Signed-off-by: Chris Diamand <chris.diamand@arm.com>
Signed-off-by: David Kilroy <david.kilroy@arm.com>