Initial Soong plugin implementation
Add an empty Soong plugin and corresponding Bob backend, as well as
simple bootstrap scripts.

This allows Android builds to run without errors (although obviously
also without building any modules), because each module now has a valid
but completely empty GenerateBuildActions implementation.

Note that `mm` will still fail, because it requires at least one module
to be defined in the directory in which it is invoked.

Currently, no configuration is done - instead, a pre-generated
`config.json` must be copied to `external/bob-build/tests/`. This will
be improved in a subsequent commit.

Change-Id: I6f67408f11fa4f45c299c0eb268fe7cb327f2e49
Signed-off-by: Chris Diamand <chris.diamand@arm.com>