Move re-usable bootstrap steps into scripts
Move the symlink creation and bootstrap file generation code into
re-usable functions in `bootstrap/utils.bash`, so that they can be used
by the Soong plugin bootstrap in a subsequent commit.

A new `bootstrap/` directory is added. In future, a few
bootstrap-related files will be moved there.

Remove `PREBUILTOS` from `bob.bootstrap.in`, because it is not
referenced or set anywhere, allowing the use of `set -u` to detect the
usage of undefined variables, which the Soong bootstrap script will
require when it uses these helpers.

Change-Id: I08112a1d344c7bec4df1331d3726c5f2a7f4d4a3
Signed-off-by: Chris Diamand <chris.diamand@arm.com>