Run config during Soong plugin bootstrap
Run the config when the Soong plugin is bootstrapped. For now, this
doesn't include any argument parsing or the ability to run menuconfig
from the bootstrap script.

Add an Mconfig option for the Soong builder, and check it when the
soongGenerator is created.

While modifying that function, remove a TODO which was implemented in
`087394f Pass config file location to Soong plugin`.

Change-Id: I2a22607889b419772125307882795e997336c7d2
Signed-off-by: Chris Diamand <chris.diamand@arm.com>