Fix exporting of external include dirs
Add a test case for the issue where re-exported libraries' include
directories were not being propagated to users. This was caused by not
specifying `LOCAL_EXPORT_[SHARED|STATIC|HEADER]_LIBRARIES` in the
generated makefiles.

Note that there is no `LOCAL_EXPORT_[...]_CFLAGS` equivalent in
Android.mk, because all exported cflags are propagated as far as the
final linking module. This is why `android_make.go` _appears_ to be only
re-exporting the include paths, rather than all cflags.

Add a check in a backend-independent mutator to verify that modules only
re-export libraries that they are actually using.

Change-Id: I309b7077321adae1d0e858a1e3f6fb3d15102c2d
Signed-off-by: Chris Diamand <chris.diamand@arm.com>