Improve match_srcs test
Setup the test so that it can verify the concatenation of the files
works as expected, and require the content of both files to be
present. Do this by compiling the output.

Make the dynamic-list option actually do something by giving it a
valid function to export. Rename the files to reflect what they would
actually be doing in this test.

Change-Id: I38e984264227cad7b4ed5ae042973eb8200226e9
Signed-off-by: David Kilroy <david.kilroy@arm.com>