Allow Bob templates and features to run in Soong LoadHook
For each Bob module we expect to instantiate an equivalent Soong
module, filling in the relevant Soong properties from the Bob
properties. In order to do this, we want to expand templates and
flatten features.

In Bob, templates and features are handled by mutators. In Soong, we
will need to do this in a LoadHook (which is called from a mutator).

Make the core Bob template and feature code agnostic as to whether
they are running in a mutator or LoadHook.

Setup an interface to identify Bob modules that need to use a soong
LoadHook, and use it for generateSource.

Change-Id: If1ab57c64b8e8fde8a3f2c1bbbcb4173fd9fce58
Signed-off-by: David Kilroy <david.kilroy@arm.com>