In Soong, override module names for Bob modules
Soong won't be outputting anything for the Bob modules. Instead we
will create equivalent Soong modules with the normal name. Therefore
rename the Bob modules to avoid conflicts.

Move the use of blueprint.SimpleName into moduleBase, so we can easily
override Name() in soong_plugin.go.

Change a number of current callers to m.Name() to explicitly call
SimpleName.Name(). Specifically change the outputs(), shortName() and
altName() functions to do this, as we want the correct output names
even if we're the Bob version in soong.

Note that there are a number of mutators that use m.Name() during
dependency walks which would be broken. These can't be run on soong.

Change-Id: I52a0ee23cd9deed8bc22e8047c1bdb7bf7c06f80
Signed-off-by: David Kilroy <david.kilroy@arm.com>