Implement bob_transform_source LoadHook
bob_transform_source maps to gensrcs. However gensrcs only allows a
new output extension to be specified instead of applying an arbitrary
regular expression to the input filename. For the moment we just
specify an extension derived from the replacement expression.
Eventually we may need a custom module type to deal with this.

Change-Id: I3e15ce978ec0990c04109b5d7c0f4a0ab4b70acc
Signed-off-by: David Kilroy <david.kilroy@arm.com>