Make menuconfig arguments available to plugins
In commit e53161f277b we modified the prototype of init_config() to
record the filename that would be used to save the config file, for
use by a plugin. This broke some external code that is also calling
init_config().

It doesn't actually make sense to pass a filename in this call, so
come up with an alternative way to get the filename.

Provide a wrapper for parse_args(), so that we can keep a copy of all
the arguments for any invocation. Provide wrapper functions to get the
appropriate options. Remove some custom code that was doing something
similar for the database directory.

Change-Id: I0dfb052986581a4aff61d948554607737d49af68
Signed-off-by: David Kilroy <david.kilroy@arm.com>