Move JSON write to main update_config
In commit e53161f we modified the prototype of init_config() to record
the filename that would be used to save the config file, for use by a
plugin. This broke some external code that is also calling
init_config().

It doesn't actually make sense to pass a filename in this call, so
come up with an alternative.

The filename is being used to write an extra JSON output. The problem
with this is that the plugin executes before the main output file is
written. Move the JSON write into the main part of the config system,
just keep the code in a separate Bob module that we only import if
needed.

Update the bootstrap version as this commit drops a plugin

Change-Id: I4692a98aa652376ce1f4fb090e552e5a5cf78da3
Signed-off-by: David Kilroy <david.kilroy@arm.com>