Run simple mutators
Run the default deps, feature and template appliers,
check_lib_fields, and strip_empty_components mutators in Soong.

This is fairly straightforward to implement - a new interface is added,
containing a subset of the Blueprint and Android mutator context
methods, which allows the existing code to be re-used.

Some methods have different signatures in each interface (e.g.
TopDownMutatorContext.Module()), mostly because of using
`android.Module` vs `blueprint.Module`. These are left out of the common
interface, and instead will be implemented using a
Soong-or-standalone-specific function. For now, only `Module()` has
this.

Add `.Parallel()` to the new and existing mutators, matching what
standalone Bob is doing.

Change-Id: I38c94848be99151dd283bf412f72075a5c5fa91a
Signed-off-by: Chris Diamand <chris.diamand@arm.com>