Rename Bob modules before creating Soong equivalents
This reverts parts of commit `69dd423 In Soong, override module names
for Bob modules`, which caused Bob modules to have a suffix on the
module names.

Instead, use the TopDownMutator's `Rename` method to rename the Bob
module at the last minute. This affects the name returned by
`mctx.ModuleName()`, but not what is returned by the `Name()` method. As
a result, we no longer need to override `Name()`.

This means that in Soong we can refer to Bob modules using their normal
names, rather than adding the special suffix each time. This will make
adding a dependency mutator in a future commit simpler.

Simplify calls to `m.SimpleName.Name()` by removing the `SimpleName`
part, as it is no longer required, as all Name() methods return the same
result.

Change-Id: Ibdfe8931444beb243e269c379801b4b8506531a0
Signed-off-by: Chris Diamand <chris.diamand@arm.com>