Run more mutators
Run the splitter, target, default applier, and depender mutators, as
well as the ones required for `reexport_libs` and the static library
graph orderer.

The only non-"mechanical" change in this is to continue to only build
target libraries - this requires fixing in setupCcLibraryProps to check
the actual variant, rather than the list of supported ones. This is
required now that we are running the splitter mutator.

The `export_lib_flags` mutator is *not* included, because it is only
used to export `ld[flags|libs]`. This should be handled by Soong.

Change-Id: I957d73eb37fad0b9c5818424b2f187ef285e5b14
Signed-off-by: Chris Diamand <chris.diamand@arm.com>