Implement exported cflags
Re-implement the getExportedVariables function for Soong (the current
implementation uses a ModuleContext, which is not _quite_ compatible
with TopDownMutatorContext, so cannot be re-used).

Update the structure of the non-Soong version for consistency.

Set ldflags and exported include dirs in the generated cc_library
module. This means that most tests based on static libraries and
binaries can be enabled for Soong.

Change-Id: I0bd866b66ba77ebe41da057b66378f4e6fc6ac68
Signed-off-by: Chris Diamand <chris.diamand@arm.com>