Build host libraries
Enable building of host libraries. This is non-trivial, because the
obvious solution (just changing the factories passed to `CreateModule`)
would create multiple modules of the same name, which does not work.

Work around that by using the library's `shortName`, which disambiguates
between the host and target versions, then fixing up the generated
filename using Soong's `stem` field.

Disable the `generate_libs` tests; generated libraries do not
currently work, because they don't set the variation correctly and
therefore cannot be linked to. This is exposed now that we support host
binaries, which are used in those tests.

Change-Id: Iade3785877595559fc121bb44cce29ca8a11ebde
Signed-off-by: Chris Diamand <chris.diamand@arm.com>