Run generated depender mutator
Fix how `Host_bin` is used in the generated build actions. Because it
refers to a module, it may also include a `:host` or `:target` suffix,
to disambiguate it. However this means that the it can never be used
directly as a string - we must instead always go via the dependency tag
added by `parseAndAddVariationDeps()`.

Make the splitter mutator name consistent with the variation name -
without this, Blueprint will not be able to find modules with matching
tags.

Adding the above fixes allows the generated depender mutator to be run.

Change-Id: Ibe1cbd6dde35f4e0fbbe8205fdc04fa1b1e50317
Signed-off-by: Chris Diamand <chris.diamand@arm.com>