config_system: support '+' and '-' operators
Add support to the lexer and grammar. In the grammar, current
expressions are all boolean. Rename 'expr' to 'condexpr'. Use 'expr'
for non-conditional expressions i.e. + and -. Ensure that choice
defaults must name an identifier, and not use an expression.

Treat booleans independently from config identifiers. Currently 'y' is
identified as a word in the grammar, and this means the higher level
logic identifies 'y' and 'n' as dependencies. By matching them to a
separate token we can maintain their separation.

With booleans separate, treat identifiers ('word') in a consistent
manner. This simplifies the logic that operates on expressions, as an
expression must always be a 2-tuple or 3-tuple.

Rename 'val' within 'default_cond' to 'expr'. This ensures we catch
most of the cases where the expression must be evaluated.

Rename value_of() to condexpr_value() since it is primarily handling
the conditional expressions (which can include + and -). Implement a
separate expr_value() to just deal with + and - in contexts where we
don't expect a conditional expression.

Reverse dependencies must be added for default values (normal and
conditional). In the conditional case, the dependency is added for all
values (not just the current).

Change-Id: Ie6b0fffbb493987ac739964e5fb52f12beba10c0
Signed-off-by: David Kilroy <david.kilroy@arm.com>