Fix exclude_srcs with relative build dir
Commit `14eda8b Run Android.mk generation in ANDROID_BUILD_TOP` added
the source dir to the inputs of the glob function. However, it did not
do the same for the excludes list.

This means that excludes are sometimes missed. Fix it by prepending the
path to the source directory before evaluating the glob.

Change-Id: I8b1557d1f984b690cc519911c18adf1c47fb5b28
Signed-off-by: Chris Diamand <chris.diamand@arm.com>