Allow abstraction of dependency-visiting helpers
Create a new abstracted context, VisitableModuleContext, representing
the contexts which support visiting of dependencies, i.e.
TopDownMutatorContext and ModuleContext.

This allows functions like generateKbuildArgs() to be called both from
GenerateBuildActions(), which has a ModuleContext, and Soong's
buildActionsMutator, which is a top-down mutator.

The abstracted interface uses private interfaces to match with the
Blueprint and Soong versions of each method. As a result,
`android_make.go` and `linux.go` are removed from the Soong build,
because they use Blueprint contexts directly.

Change-Id: Ibf4ad06c30fe8028eedf5fa150b4fd94adb05a76
Signed-off-by: Chris Diamand <chris.diamand@arm.com>